Make quantity selector support decimals

+++ app/src/main/java/io/github/sheasmith/discountdetective/ui/components/search/SearchProductCard.kt
@@ -53,7 +53,7 @@ fun SearchProductCard(
     retailers: Map<String, Retailer>,
     snackbarHostState: SnackbarHostState,
     coroutineScope: CoroutineScope,
-    onAddToShoppingList: ((String, String, String, Int) -> Unit)? = null,
+    onAddToShoppingList: ((String, String, String, Double) -> Unit)? = null,
     region: String
 ) {
     val product = productPair?.second
